trixnity 3.7.0

+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/i18n/I18n.kt
@@ -94,9 +94,9 @@ abstract class I18n(languages: Languages, settings: MatrixMessengerSettingsHolde
         DE - "Gruppe"
     }
 
-    fun roomNameOther(othersCount: Long) = translate {
-        EN - if (othersCount == 1L) "one other" else "$othersCount others"
-        DE - if (othersCount == 1L) "ein anderer" else "$othersCount andere"
+    fun roomNameOther(othersCount: Int) = translate {
+        EN - if (othersCount == 1) "one other" else "$othersCount others"
+        DE - if (othersCount == 1) "ein anderer" else "$othersCount andere"
     }
 
     fun eventChangeAvatar(username: String) = translate {
